reorganize menu components
- move profile menu componets to its own file
