+++ package.json
@@ -26,12 +26,17 @@
     "husky": "^3.0.7",
     "lint-staged": "^9.4.0",
     "prettier": "^1.18.2",
+    "stylelint": "^11.1.1",
+    "stylelint-config-recommended": "^3.0.0",
+    "stylelint-config-styled-components": "^0.1.1",
+    "stylelint-processor-styled-components": "^1.8.0",
     "typescript": "^3.6.3",
     "typescript-plugin-styled-components": "^1.4.3"
   },
   "scripts": {
     "start": "react-app-rewired start",
     "build": "react-app-rewired build",
+    "lint:css": "stylelint './src/**/*.tsx'",
     "test": "react-app-rewired test",
     "eject": "react-scripts eject"
   },
