+++ src/Components/logo/SidebarMenuLogo/SidebarMenuLogo.tsx
@@ -17,7 +17,7 @@ const SidebarMenuLogo: React.FunctionComponent<LogoProps> = props => {
 
   if (type === 'small') {
     return (
-      <MenuLogo
+      <StyledMenuLogo
         {...localProps}
         src="/img/sidebarmenu-logo-74x45.jpg"
         alt="collapsed application logo"
@@ -26,7 +26,7 @@ const SidebarMenuLogo: React.FunctionComponent<LogoProps> = props => {
   }
 
   return (
-    <MenuLogo
+    <StyledMenuLogo
       {...props}
       src="/img/sidebarmenu-logo-192x45.jpg"
       alt="expanded application logo"
@@ -34,7 +34,7 @@ const SidebarMenuLogo: React.FunctionComponent<LogoProps> = props => {
   );
 };
 
-const MenuLogo = styled('img')`
+const StyledMenuLogo = styled('img')`
   height: 32px;
   margin: 16px 0 16px 8px;
   background: #fff;
