+++ src/Components/menu/ProfileMenu/index.ts
@@ -1,0 +1,1 @@
+export { default as profileMenu } from './ProfileMenu';
